requirements.txt
Django==5.0.6
djangorestframework==3.15.2
jwt==1.3.1
google-auth==2.32.0
google==1.9.3
auth==0.5.3
PyJWT==2.8.0
psycopg2==2.9.9